requirements.txt
langchain
langchain-groq
langchain_community
python-dotenv
streamlit
pandas
PyMuPDF
python-docx
docx2pdf